data_science/src/requirements.txt
bcrypt
cfenv
Faker
fastapi
Flask-WTF
gunicorn
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
Jinja2
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-psycopg2
passlib
psycopg2-binary
python-jose
scikit-learn
spaCy
uvicorn
sympy
python-multipart
python-dotenv